refactor(xpath): split xpath.c into ast/lexer/parser/eval/functions
One translation unit became six files behind a shared internal.h: the arena
AST/dump (ast.c), tokenizer (lexer.c), recursive-descent parser plus the
//X optimization (parser.c), the evaluator and value model (eval.c), and the
function library with EXSLT regex and the Python test hook (functions.c).
No behavior change; the engine's public surface in xpath.h is untouched.
