+++ meson.build
@@ -32,7 +32,11 @@ py.extension_module(
         'src/turbohtml/_c/serialize/escape.c',
         'src/turbohtml/_c/serialize/unescape.c',
         'src/turbohtml/_c/serialize/markup.c',
-        'src/turbohtml/_c/query/xpath/xpath.c',
+        'src/turbohtml/_c/query/xpath/ast.c',
+        'src/turbohtml/_c/query/xpath/lexer.c',
+        'src/turbohtml/_c/query/xpath/parser.c',
+        'src/turbohtml/_c/query/xpath/eval.c',
+        'src/turbohtml/_c/query/xpath/functions.c',
         'src/turbohtml/_c/features/sanitize.c',
         'src/turbohtml/_c/features/linkify.c',
         'src/turbohtml/_c/features/links.c',
